pulsar_bank

+++ src/main/java/com/example/wallet/DTO/CurrencyDTO.java
@@ -7,18 +7,20 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class CurrencyDTO {
     @Min(value = 1, message = "Айди не может быть пустым или нулевым")
-    private int id;
+    private Long id;
 
     @NotNull(message = "Выберите операцию")
     private String operation;
 
     @NotNull(message = "Сумма не должна быть пустой")
     @DecimalMin(value = "0.01", message = "Сумма не должна быть пустой или нулевой")
-    private Double amount;
+    private BigDecimal amount;
 
 }
