+++ src/main/java/com/example/wallet/DTO/FillDTO.java
@@ -7,17 +7,19 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class FillDTO {
     @NotNull(message = "Айди не может быть пустым")
     @Min(value = 1, message = "Айди не может быть пустым или быть меньше 1")
-    private Integer id;
+    private Long id;
 
     @NotNull(message = "Сумма не должна быть пустой")
     @DecimalMin(value = "0.01", message = "Сумма не должна быть меньше 0.01 или пустой")
-    private Double amount;
+    private BigDecimal amount;
 
     @NotNull(message = " - Выберите валюту")
     private String currency;
